Convert current and temperature to int and float

+++ sensors/tasks.py
@@ -1,11 +1,9 @@
-from __future__ import absolute_import, unicode_literals
-
 from celery import shared_task
 from serial import Serial
 
 
 @shared_task
-def read(arduino: str) -> (float, float):
+def read(arduino: str) -> (int, float):
     """
     Does a single read on the serial interface to the Arduino reading the
     current power and temperature.
@@ -14,4 +12,4 @@ def read(arduino: str) -> (float, float):
     with Serial(arduino) as ser:
         read_arduino = ser.readline().decode("utf-8").strip()
         current, temp = read_arduino.split("\t")
-    return current, temp
+    return int(current), float(temp)
